requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
streamlit==1.26.0
langchain==0.0.312
langchain_openai==0.0.2
supabase==1.0.3
openai>=1.6.1
python-dotenv==1.0.0
pypdf==3.16.0